requirements.txt
altgraph==0.17.4
astroid==3.3.5
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
cssutils==2.11.1
DateTime==5.5
dill==0.3.9
disposable-email-domains==0.0.107
dnspython==2.6.1
idna==3.10
isort==5.13.2
lxml==5.3.0
macholib==1.16.3
mccabe==0.7.0
more-itertools==10.5.0
packaging==24.1
platformdirs==4.3.6
premailer==3.10.0
pyfiglet==1.0.2
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pylint==3.3.1
PySocks==1.7.1
pytz==2024.2
requests==2.32.3
setuptools==75.1.0
tomlkit==0.13.2
tzlocal==5.2
urllib3==2.2.3
yagmail==0.15.293
zope.interface==7.0.3
